bi_analytics/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
mysql-connector-python==9.1.0
folium==0.15.1
numpy==1.26.4
pandas==2.2.0
PyYAML==6.0.1
python-dotenv==1.0.1
pydantic==2.5.3
jinja2==3.1.3
